+++ uint256.h
@@ -12,8 +12,8 @@
 #include <string>
 #include <vector>
 
-typedef long long  int64;
-typedef unsigned long long  uint64;
+typedef int64_t int64;
+typedef uint64_t uint64;
 
 
 inline int Testuint256AdHoc(std::vector<std::string> vArg);
